% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online2SpamH-package.R
\docType{package}
\name{Online2SpamH-package}
\alias{Online2SpamH}
\alias{Online2SpamH-package}
\title{Online2SpamH: Apply Online 2SpamH algorithm}
\description{
The package applies the 2SpamH algorithm to the passively collected variable and assigns it a "Missing" or "Non-missing" label based on auxiliary device usage and sensor activity information.
}
\author{
\strong{Maintainer}: Sebastian Brylka \email{seb4012@med.cornell.edu}

}
\keyword{internal}
