% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_2spamh.R
\name{Online_TwoSpamH}
\alias{Online_TwoSpamH}
\title{Online_TwoSpamH}
\usage{
Online_TwoSpamH(
  new_data,
  training_data,
  passive_variable,
  phone_usage_vars,
  activity_level_vars,
  num.neighbor = 5,
  seed = NULL,
  plot.data = F
)
}
\arguments{
\item{new_data}{A data frame with a single new observation to be labeled as "Missing" or "Non-missing".}

\item{training_data}{A data frame processed by a 'TwoSpamH_train()' function.}

\item{passive_variable}{A passive variable name to be labeled as "Missing" or "Non-missing".}

\item{phone_usage_vars}{A vector of strings of phone usage variable names from data.}

\item{activity_level_vars}{A vector of strings of activity level variable names from data.}

\item{num.neighbor}{A number of neighbors considered by each initially unlabeled data point in KNN algorithm.}

\item{seed}{The seed to be set, default is NULL.}

\item{plot.data}{Whether to plot the data with labeled and highlighted new observation, default is FALSE.}
}
\value{
A new_data data frame with "Missing"/"Non-missing" label assignment.
}
\description{
A function to label a new observation as "Missing" or "Non missing" using a training set trained with 'TwoSpamH_train()' function.
}
\examples{
# Load example data
data("example_data")

# Define variable names
passive_variable <- "step_count"
activity_level_vars <- "n_uploads"
phone_usage_vars <- c("screen_unlocks", "display_events")

# Train the model
train <- TwoSpamH_train(
  data = example_data,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  plot.data = FALSE
)

# Create a new observation
new_data <- data.frame(
  step_count = 3400,
  n_uploads = 90,
  screen_unlocks = 25,
  display_events = 130
)

# Predict using the trained model
Online_TwoSpamH(
  new_data = new_data,
  training_data = train,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  plot.data = TRUE
)
}
