% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_2spamh.R
\name{TwoSpamH}
\alias{TwoSpamH}
\title{TwoSpamH}
\usage{
TwoSpamH(
  data,
  passive_variable,
  phone_usage_vars,
  activity_level_vars,
  thresholds = data.frame(lower_bound_phone_usage = 0.3, upper_bound_phone_usage = 0.7,
    lower_bound_activity_level = 0.3, upper_bound_activity_level = 0.7),
  num.neighbor = 5,
  check.cor = NULL,
  plot.data = F,
  seed = NULL
)
}
\arguments{
\item{data}{A data frame containing the variable to be labelled and the variables representing device usage and sensory activity level.}

\item{passive_variable}{A passive variable name to be labeled as "Missing" or "Non-missing".}

\item{phone_usage_vars}{A vector of strings of phone usage variable names from data.}

\item{activity_level_vars}{A vector of strings of activity level variable names from data.}

\item{thresholds}{A data frame with lower and upper thresholds for device usage and sensor activity level (look at examples for formatting).}

\item{num.neighbor}{A number of neighbors considered by each initially unlabeled data point in KNN algorithm.}

\item{check.cor}{Whether to remove one of the device usage or sensor activity vectors if they are highly correlated. If no, input should be NULL (default). If yes, input should be the correlation threshold for variables to be removed.}

\item{plot.data}{Whether to plot the data, default is FALSE.}

\item{seed}{The seed to be set, default is NULL.}
}
\value{
The original input data frame with an additional 'label' column representing observation's "Missing"/"Non missing" status.
}
\description{
A function to run 2SpamH algorithm that assigns all observations in data frame a "Missing" or "Non-missing" label using a pre-defined set of lower and upper thresholds.
}
\examples{
# Load example data
data("example_data")

# Define variables
passive_variable <- "step_count"
activity_level_vars <- "n_uploads"
phone_usage_vars <- c("screen_unlocks", "display_events")
all_vars <- c(passive_variable, activity_level_vars, phone_usage_vars)

# Run TwoSpamH with example thresholds
thresholds <- data.frame(
  lower_bound_phone_usage = 0.2,
  upper_bound_phone_usage = 0.8,
  lower_bound_activity_level = 0.3,
  upper_bound_activity_level = 0.7
)

TwoSpamH(
  data = example_data,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  thresholds = thresholds,
  plot.data = TRUE
)

}
