% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_2spamh.R
\name{TwoSpamH_train}
\alias{TwoSpamH_train}
\title{TwoSpamH_train}
\usage{
TwoSpamH_train(
  data,
  passive_variable,
  phone_usage_vars,
  activity_level_vars,
  method = "majority_vote",
  num.neighbor = 5,
  seed = NULL,
  plot.data = F,
  conf = F
)
}
\arguments{
\item{data}{A data frame containing the variable to be labelled and the variables representing device usage and sensory activity level.}

\item{passive_variable}{A passive variable name to be labeled as missing or non-missing.}

\item{phone_usage_vars}{A vector of strings of phone usage variable names from data.}

\item{activity_level_vars}{A vector of strings of activity level variable names from data.}

\item{method}{Which Online2SpamH method to use. Choose from "majority_vote" (default) or "optimal_thresholds".}

\item{num.neighbor}{A number of neighbors considered by each initially unlabeled data point in KNN algorithm.}

\item{seed}{The seed to be set, default is NULL.}

\item{plot.data}{Whether to plot the data, default is FALSE.}

\item{conf}{Whether to calculate confidence of label assignment, append it to original data frame, default is FALSE.}
}
\value{
The original input data frame with an additional column 'label' representing observation's "Missing"/"Non missing" status.
}
\description{
A function to train the model by assigning "Missing"/"Non missing" labels to all available observations. By default, it uses a Majority Vote method - check manuscript for details.
}
\examples{
# Load example data
data("example_data")

# Define variable names
passive_variable <- "step_count"
activity_level_vars <- "n_uploads"
phone_usage_vars <- c("screen_unlocks", "display_events")
all_vars <- c(passive_variable, activity_level_vars, phone_usage_vars)

# Training function
TwoSpamH_train(
  data = example_data,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  method = "majority_vote",
  plot.data = TRUE
)

}
