% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Step Count Dataset}
\format{
A data frame with \code{n} rows and the following columns:
\describe{
\item{step_count}{Number of steps per day}
\item{n_uploads}{Number of sensor uploads to the server}
\item{screen_unlocks}{Number of phone unlocks}
\item{display_events}{Number of display events}
}
}
\usage{
data(example_data)
}
\description{
A dataset containing passive sensing features from the Alacrity study.
}
\keyword{datasets}
