% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Online2SpamH-package.R
\docType{package}
\name{Online2SpamH-package}
\alias{Online2SpamH}
\alias{Online2SpamH-package}
\title{Online2SpamH: Apply Online 2SpamH algorithm}
\description{
A two-stage algorithm for processing passively sensed mHealth data that uses auxiliary device usage and sensor activity information to determine device-use status, with non-use classified as “Missing” and use as “Non-Missing”.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sambanlab/Online2SpamH}
  \item Report bugs at \url{https://github.com/sambanlab/Online2SpamH/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Brylka \email{seb4012@med.cornell.edu}

}
\keyword{internal}
