% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_2spamh.R
\name{Online_TwoSpamH}
\alias{Online_TwoSpamH}
\title{Online_TwoSpamH}
\usage{
Online_TwoSpamH(
  new_data,
  training_data,
  passive_variable,
  phone_usage_vars,
  activity_level_vars,
  num.neighbor = 5,
  seed = NULL,
  plot.data = F
)
}
\arguments{
\item{new_data}{new observation to be labeled as "Missing" or "Non-missing" based on the training data}

\item{training_data}{training dataset after TwoSpamH_train() function}

\item{passive_variable}{name of variable to be marked as missing or non-missing}

\item{phone_usage_vars}{vector of strings of phone usage variable names from data}

\item{activity_level_vars}{vector of strings of activity level variable names from data}

\item{num.neighbor}{neighbors for KNN algorithm}

\item{seed}{set seed}

\item{plot.data}{whether to plot a data with new observation}
}
\value{
new_data observation with "Missing"/"Non-missing" label assignment
}
\description{
Label a new observation as "Missing" or "Non missing" using a training set
}
\examples{
# Load example data
data("example_data")

# Define variable names
passive_variable <- "step_count"
activity_level_vars <- "n_uploads"
phone_usage_vars <- c("screen_unlocks", "display_events")

# Train the model
train <- TwoSpamH_train(
  data = example_data,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  plot.data = FALSE
)

# Create a new observation
new_data <- data.frame(
  step_count = 3400,
  n_uploads = 90,
  screen_unlocks = 25,
  display_events = 130
)

# Predict using the trained model
Online_TwoSpamH(
  new_data = new_data,
  training_data = train,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  plot.data = TRUE
)
}
