% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_2spamh.R
\name{TwoSpamH_train}
\alias{TwoSpamH_train}
\title{TwoSpamH_train}
\usage{
TwoSpamH_train(
  data,
  passive_variable,
  phone_usage_vars,
  activity_level_vars,
  num.neighbor = 5,
  seed = NULL,
  plot.data = F,
  conf = F
)
}
\arguments{
\item{data}{data frame to run 2SpamH on}

\item{passive_variable}{name of variable to be marked as missing or non-missing}

\item{phone_usage_vars}{vector of strings of phone usage variable names from data}

\item{activity_level_vars}{vector of strings of activity level variable names from data}

\item{num.neighbor}{neighbors for KNN algorithm}

\item{seed}{set seed}

\item{plot.data}{whether to plot data}

\item{conf}{if plot.data = TRUE, whether to plot the confidence of point's assignment}
}
\value{
original dataset with column 'label' with "Missing"/"Non-missing" values
}
\description{
A function to train a dataset by assigning 2SpamH labels to all observations using a Majority Vote method
}
\examples{
# Load example data
data("example_data")

# Define variable names
passive_variable <- "step_count"
activity_level_vars <- "n_uploads"
phone_usage_vars <- c("screen_unlocks", "display_events")
all_vars <- c(passive_variable, activity_level_vars, phone_usage_vars)

# Training function
TwoSpamH_train(
  data = example_data,
  passive_variable = passive_variable,
  phone_usage_vars = phone_usage_vars,
  activity_level_vars = activity_level_vars,
  plot.data = TRUE
)

}
